Add capability to match on device number
Irrespective of the USB device path (which is only available for certain
versions of libusb), one could use the USB device number (as shown in
--list) to identify the USB device uniquely in case multiple devices of
the same vendor/product ID are attached.

The suggestion of the original code to unplug one device doesn't work in
case of composite USB devices which contain multiple hard-wired devices
with the same vendor/product ID.  In this case it's impossible to unplug
one of them.

Signed-off-by: Harald Welte <laforge@gnumonks.org>