Get GD32VF103 serial number as UTF-8/ASCII
This device violates the USB specification and doesn't encode its serial
number as UTF-16. If the serial number were just a serial number, it'd
be fine to just print some '?'s like we do now. However, the serial
number field on this device is used to hold the part number, and we need
access to that in order to work around a different bug.

Add a quirk that allows a device's serial number to be read as raw bytes
instead of as UTF-16 codepoints, and enable it for GD32VF103.

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>