Override reported GD32VF103 layout
This chip has a bootloader bug that makes it return the wrong page size
and page count for its internal flash. While this is merely an annoyance
for uploads (we get extra data), it breaks downloads completely since
only half the pages are erased and written. (Every other page is left
untouched when we use the reported page size of 2048.)

Nucleisys (the core designer) has a fork of dfu-utils with a fix for
this bug [1]. and this patch is a cleaned-up version of that. I've added
a new quirk that allows custom fixup code to run after the memory layout
is read but before it's used, so much of this code is applicable to
other devices that may have similar bugs.

[1]: https://github.com/riscv-mcu/gd32-dfu-utils/commit/c2c8215061b08146e6a4c8c22bc57fae95c656df

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>