man: Generate HTML using pandoc
This produces nice looking output with the stock settings, meaning
there's no need to manually maintain the HTML version anymore.

Since pandoc has a ton of Haskell dependencies, I'm keeping the
generated HTML tracked in Git, but I've also added a Makefile rule so it
can easily be regenerated by people who change the groff documentation.

I also added a section of explicit indentation to the groff file to
avoid a quirk of Pandoc's rendering that caused the paragraph starting
with "If only a single serial number is specified" to not be properly
indented in the HTML version.

Signed-off-by: Thomas Hebb <tommyhebb@gmail.com>