dfuse-pack.py: Allow multiple alternate interfaces
For binary files, an alternate interface number can be
specified for each (DfuSe "image element"):

    -b address@altintf:file

If none is specified, the default is used (0 or specified by
the -a option).

If two consecutive image elements have the same alternate
interface number, they are packed into the same DfuSe "DFU image".

As before, binary files and hex files (will have default alternate
interface number) can be combined, but not s19 files.

FIXME: Internally the alternate interface number is part of every
element handed over to build(), ideally it should be per DFU image
("target" in the current code), but this works for now.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>