dfuse: Warn about future change in DfuSe file handling
The change will be backwards compatible but a little
heads-up might be good.

The plan is that, like today, if -a X is used to specify
one alternate interfaces, only matching DFU images
(parts of the DfuSe file targeting one alternate interface)
will be downloaded.

In the future, if -a is not used, it will automatically
select the alternate interfaces and download all the DFU
images in the DfuSe file. Today it will stop and ask the
user to specify one alternate interface.

https://sourceforge.net/p/dfu-util/tickets/34/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>