dfuse: Improve progress bar and verbose output
The page erase was originally a sub-step of the page write.

Since erase and write are done in separate passes, leave the erase pass
progress bar completed and start on a new line for the writing pass.

If verbose, do not print any progress bar elements, and print the erase
operation information as we do for writes.

Print poll timeout only if -v -v.

Also direct some error messages to stderr.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>