dfuse: Workaround for STM32L4 page erase stalling USB pipe
The ROM bootloader on some STM32F4 devices reports a too short poll
timeout, so it is not finished when we poll again. Most of the time it
will report DFU_STATE_dfuDNBUSY and everything is fine after the second
poll, but sometimes, seemingly randomly, it will stall the USB pipe. For
a limited number of retries we will therefore treat this the same as a
DFU_STATE_dfuDNBUSY, reusing the last reported poll timeout.

Tested on a STM32L433RCT6.

Thanks to Rob Riggs for reporting and investigation, and to David
Flogeras for help with testing.

https://sourceforge.net/p/dfu-util/tickets/40/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>