Skip setting alternate interface when clearly not needed
Keep setting alternate interface if the DFU interface is not the
first interface. This seems to be required for some devices, see
commit 3193a111.

Otherwise only set alternate interface if the interface actually
have multiple alternate interfaces.

This is to avoid setting it when it is not needed, since the request is
not supported by all devices. In most cases the OS kernel or libusb
takes care of a resulting pipe error, but this is not consistent among
the platforms, and is currently not working well on macOS (should be
improved in libusb 1.0.25).

There is a possibility that some other devices need this transaction
for more obscure reasons, but these would need to be identified and
dealt with through quirks.

https://sourceforge.net/p/dfu-util/tickets/86/

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>