Improve messages about Run-Time status
Some devices are already in DFU mode instead of the Run-Time mode that
we initially expect.

Report the state from first GET_STATUS if not in Run-Time mode.

Also spell Run-Time consistently in our output, capitalized to
underline we are meaning the Run-Time versus DFU mode from the
DFU specification.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>